修正删除日志不记录历史问题

+++ osight-web/src/main/webapp/article/list.jsp
@@ -3,7 +3,7 @@
 <%@ taglib prefix="s" uri="/struts-tags"%>
 
 <div>
-	<s:a namespace="/article" action="add" cssClass="button ui-button-success">撰写新日志</s:a>
+	<s:a href="/article/new"  cssClass="button ui-button-success">撰写新日志</s:a>
 </div>
 <s:if test="page.list==null">
 还没有撰写过日志！
@@ -25,7 +25,7 @@
 							value="pv" /></span>
 							<span class="comments">
 					<s:a href="/article/edit/%{id}">修改</s:a>
-					<s:a href="/article/delete/%{id}">删除</s:a></span>
+					<s:a href="/article/delete/%{id}" onclick="return confirm('确认要删除吗？')">删除</s:a></span>
 				</div>
 				<div class="content">
 					<s:property value="content" escapeHtml="false"></s:property>
